requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
torchaudio
jiwer
sentence-transformers
transformers
scikit-learn
language-tool-python
streamlit
requests
SpeechRecognition
pydub